USB: Port mspdebug to newer libusb-1.0 API [WIP]
*** THIS BRANCH IS VOLATILE (Git Rebase by Force) ***

libusb-1.0 is now available on most platforms. Port utilities and
transports to the newer API. This patch will be a complete switch to
libusb-1.0 with no selection of the older API. This should keep the code
very close to the original design intent.

Files to be modified:

[wip] Makefile
[wip] util/usbutil.c
[wip] transport/cp210x.c
[wip] transport/cdc_acm.c
[wip] transport/ti3410.c
[wip] transport/bslhid.c
[wip] transport/ftdi.c
[wip] transport/rf2500.c

Platforms Tested:

- Linux (Arch and Ubuntu) 64-bit
- Windows 7 (Cygwin)

Boards/Interfaces:

- TI MSP430 LaunchPad [rf2500]
- TI MSP-FET430UIF [uif]